+++ sdk/storage/storage-file-datalake/CHANGELOG.md
@@ -12,6 +12,20 @@
 
 - Migrated to TypeSpec-based code generation [PR #38232](https://github.com/Azure/azure-sdk-for-js/pull/38232)
 
+## 12.31.0 (2026-06-15)
+
+### Features Added
+
+- Includes all features released in 12.31.0-beta.1.
+
+## 12.31.0-beta.1 (2026-04-29)
+
+### Features Added
+
+- Added support for service version 2026-06-06.
+- Added DataLakePathClient.getSystemProperties().
+- Added DataLakePathClient.getTags() and .setTags().
+
 ## 12.30.0 (2026-05-22)
 
 ### Features Added